move useLoading
